+++ src/pages/docs/changelogData.js
@@ -40,6 +40,7 @@ export const RELEASES = [
           'addedPlantIssuesWidget',
           'addedHoraPorHora',
           'addedSorting',
+          'addedDeleteUser',
         ],
       },
       {